refactor: user input on its own file
