+++ robots/user-input.js
@@ -1,0 +1,18 @@
+const readline = require('readline-sync')
+
+function robot(content) {
+  content.searchTerm = askAndReturnSearchTerm()
+  content.prefix = askAndReturnPrefix(content.searchTerm)
+
+  function askAndReturnSearchTerm() {
+    return readline.question('Type a Wikipedia search term: ')
+  }
+
+  function askAndReturnPrefix(searchTerm) {
+    const prefixes = ['Who is', 'What is', 'The history of']
+    const selectedPrefixIndex = readline.keyInSelect(prefixes, `Select a prefix for ${searchTerm}:`)
+    return prefixes[selectedPrefixIndex]
+  }
+}
+
+module.exports = robot
